Fix search for native artifacts to respect excluded dependencies
There was a subtle bug in the NativeHelper code (recently transplanted
from ApkMojo) - it threw away exclusion information, meaning that even if
a user had carefully excluded an unavailable transient dependency, the
build would still die when the native-artifact search attempted to examine
the dependency and found it (surprise!) unavailable.

This patch fixes the search for native artifacts in transient
dependencies by iterating over the project's Dependency objects
rather than Artifacts - and uses the RepositoryUtils.toDependency() method
to get an Aether Dependency object which retains the Exclusions data.

When I originally encountered this bug I made a few notes here:

http://code.google.com/p/maven-android-plugin/issues/detail?id=166#c8

I now realise the original reporter was possibly not suffering from
exactly the same bug (although it was also to do with exclusions).