Update to bug-fix release of android-screenshot lib
This release fixes parsing of the key-value pairs that can be passed
with log messages to the paparazzo service - these are only used for
naming the screenshot files at the moment.

Additionally, it fixes a race condition that could see the paparazzo
service being shut down at the same time as the last screenshot being
taken, thus corrupting the animated gif. Works much better now :-)