Updated Apk with the new ConfigHandler
Previous code used the ConfigHelper which was able to handle
configuration parameters deprecation.

Since the new ConfigHandler doesn't provide this mechanism - or I could
not find it - only the new "metaIncludes" parameter follows the POJO
configuration rules.

Other parameters are kept as they are and should be modified by the actual
maintainer of this class.