Support BuildConfig generation.
Create BuildConfig java source files in parallel with the R files. The boolean
value of the DEBUG field in this class is controlled by the 'android.release'
property. By default it will be 'true'. You can toggle it by including
'-Dandroid.release=true' on the command line or through a profile.