Added support to for updating android:authorities attributes on provider elements in the manifest
Because provider authorities are in the global package namespace (like package names) they must be updated when producing a multi configuration build that allows multiple configurations to be installed on the same device.

This update allows the updating of the android:authorities attribute of the <provider> element in the Android manifest file.