--non-constant-id for generating R for APKLIBs
Adds the --non-constant-id flag to aapt when generating R.class for
APKLIBs to give non-final ints. This is additional to the fix by xian
which was only doing this is in generateRForApkLibDependency method - this
change does it in the generateR method as well but only for APKLIBs.

This is necessary for when you want to make an apklib that has your source
as a .jar but also contain some XML resource files. It doesn't fix issue
313 but it basically allows you to do what the person who raised it
wanted.