It seems that the aar/classes.jar should NOT include the R class or BuildConfig.  If the R is included, dex will fail with an error about finding the same R in multiple jars.
I haven't really found anyone saying this explicity nor any docs that say as much.  I do have a gradle build that depends on an aar that is generated with this mojo that will fail if R is in the classes.jar.  I also know that the android gradle build tools combine resources.  The only other evidence I have that hints at not including R is the following discussion on google groups:

https://groups.google.com/forum/#!msg/adt-dev/uWppd9l8UNw/w6Q_9N13mLcJ