Fix release plugin for aar dependencies
When creating a system scoped depenedency do not add
"_expanded" suffix.  When the release plugin performs
the prepare goal it will check dependencies artifactId
against the parent artifact's artifactId.  When the
suffix is added this check will fail telling the
release plugin that the dependency is not part of the
reactor and will fail the build