Adding proguard support to library (AAR) projects
This commit alows obfuscated classes to be used while
building an AAR when proguard is enabled.

No dependencies are processed or included in the
classes.jar archive as that would cause potential code
duplication and affect dependency resolution.

The final application should take care of obfuscating
all depdendencies when applicable.

Signed-off-by: David Marques <dpsmarques@gmail.com>