Work around for manifest merger (v2) messing up certain CRLF end-of-line markers -- fixes #511
Problem: When fed a manifest file with Windows-style line endings (CRLF) the manifest-merger mojo replaces certain CRLF occurrences (specifically in multi-line XML comments) with Unix-style LF line endings, leaving the file with inconsistent line endings. Note: the (deprecated) manifest-update mojo exhibits the same problem.
Because the manifest-merger mojo relies on the Android builder library the problem should actually be tackled there (and we should submit a bug report to Google to that effect). But in the meantime, here is a stopgap solution...

Work around being committed: before letting the builder merge the manifest(s) the I check whether the androidManifestFile has CRLF line endings, if so I convert the file to have LF line endings everywhere. After the builder has done its work the line endings of in the destinationManifestFile are converted by to CRLF if the original  androidManifestFile had CRLF line endings.

Known issue: This commit only applies to the manifest-merger mojo, and does not correct the problem for the deprecated manifest-update mojo.

Signed-off-by: Matthias Stevens <matthias.stevens@gmail.com>