Add extra arguments at (or near) the end of command
If you want to add additional raw file directories to the apk, you can specify them in extra arguments - however, these should come last in the aapt command (with the exception of --output-text-symbols - which is used when generating the R.txt file.  This PR just moves the addExtraArguments later in the command list.