Make manifest merging less verbose
Currently, when you do manifest merging, it writes out all entries as `info` entries.  When you are storing the report to a file (using the `mergeReportFile` configuration parameter), this is extra information that doesn't need to be output to the screen (because it's in the file).  This patch adds an option to MavenILogger to treat `info` calls as `verbose`, and then uses that flag whenever the `mergeReportFile` parameter is specified.